requirements.txt
flask==2.0.1
flask-login==0.5.0
flask-socketio==5.1.1
flask-sqlalchemy==2.5.1
flask-limiter==1.4
flask-caching==1.10.1
flask-talisman==0.8.1
python-dotenv==0.19.0
sqlalchemy==1.4.23
pytest==6.2.5
requests==2.26.0
python-socketio==5.4.0
eventlet==0.33.0
bcrypt==3.2.0
gunicorn==20.1.0
flask-limiter==2.8.1
flask-talisman==1.0.0
flask-caching==2.0.2
redis==4.5.1  # Only if you plan to use Redis for caching in production
